agh i dont know anymorre honestly just leave it like this shitcode getting triggered 10 times cause its in shared prediction is probably not needed for this and i cant be bothered
