+++ Content.Goobstation.Shared/Security/ContrabandIcons/SharedContrabandIconsSystem.cs
@@ -25,11 +25,6 @@ public abstract class SharedContrabandIconsSystem : EntitySystem
     public override void Initialize()
     {
         base.Initialize();
-        SubscribeLocalEvent<VisibleContrabandComponent, DidEquipEvent>(OnEquip);
-        SubscribeLocalEvent<VisibleContrabandComponent, DidUnequipEvent>(OnUnequip);
-
-        SubscribeLocalEvent<VisibleContrabandComponent, DidEquipHandEvent>(OnEquipHands);
-        SubscribeLocalEvent<VisibleContrabandComponent, DidUnequipHandEvent>(OnUnequipHands);
         
         Subs.CVar(_configuration, GoobCVars.ContrabandIconsEnabled, value => _isEnabled = value);
     }
@@ -51,25 +46,5 @@ public abstract class SharedContrabandIconsSystem : EntitySystem
             ContrabandStatus.Contraband => "ContrabandIconContraband",
             _ => "ContrabandIconNone"
         };
-    }
-
-    private void OnEquip(EntityUid uid, VisibleContrabandComponent component, DidEquipEvent args)
-    {
-        ContrabandDetect(uid, component, args.SlotFlags);
-    }
-
-    private void OnUnequip(EntityUid uid, VisibleContrabandComponent component, DidUnequipEvent args)
-    {
-        ContrabandDetect(uid, component, args.SlotFlags);
-    }
-
-    private void OnUnequipHands(EntityUid uid, VisibleContrabandComponent component, DidUnequipHandEvent args)
-    {
-        if (!HasComp<IdCardComponent>(args.Unequipped) && uid == args.User)
-            ContrabandDetect(uid, component);
-    }
-    private void OnEquipHands(EntityUid uid, VisibleContrabandComponent component, DidEquipHandEvent args)
-    {
-            ContrabandDetect(uid, component);
     }
 }
